Rename lib folder
