--- packages/safe-kit/src/lib/index.ts
@@ -1,2 +1,0 @@
-export * from './sendTransaction'
-export * from './sendAndDeployTransaction'
